chore(benchmark): add roundtrip benchmark
