+++ package.json
@@ -20,7 +20,8 @@
     "prepare": "husky",
     "fix:lint-format": "npm run lint:fix && npm run format",
     "prepublishOnly": "npm run build && npm test",
-    "benchmark": "npm run build && npm link && npx jiti benchmark/run-benchmark.ts"
+    "benchmark": "npm run build && npm link && npx jiti benchmark/runner/run-benchmark.ts",
+    "benchmark:roundtrip": "npm run build && npm link && npx jiti benchmark/runner/run-roundtrip-benchmark.ts"
   },
   "repository": {
     "type": "git",
